Clean up Travis config
Currently the language is specified as Scala, but there seems no reason for that to be known as the app is built using a script. Travis automatically updates SBT as part of the build, but SBT is not used to build this app anymore, so hopefully this will save some time.

Also the Heroku deploy from Travis is picking up a buildpack that requires SBT, and it is not clear why this should be the case - possibly specifying Scala here has this effect.