Bugfix - serve furniture as resources from the assembled jar
Needs tidying up, code is a Java solution from StackOverflow (https://stackoverflow.com/a/1264737) but the live site is currently down, so committing and deploying in the current state.

Furniture tests and configurable furniture path are no longer necessary - tests don't make much sense when the furniture is bundled and versioned with the app anyway, it isn't so generic. Will need new tests to confirm that CSS file is served as expected, perhaps.