Start implementing a custom HTTP client to replace Dispatch
Dispatch is quite heavyweight in terms of dependencies, but other Scala HTTP clients are either similarly heavyweight or do not implement an API that I like.

The standard Java URL method works reasonably well, so this is a fairly simple Scala wrapper around that using my preferred abstractions and patterns.

Currently only supports GET requests, and as such is not yet in-use.