Start to split request / response model into separate module for reuse
The model was essentially duplicated between Cicerone and Sommelier, and a similar model is used in the DueSoldi functional tests. By modularising it in a generic way all three can benefit from changes made to it.

Ratatoskr is the squirrel which passes messages along the trunk of the world-tree in Norse mythology...