Remove the Cicerone request builder in favour of Ratatoskr
This provides a generic request builder that can be shared across projects.

The builder is separate from the data so it can be imported optionally and to separate concerns.

This has led to some tests in Sommelier becoming more verbose when creating requests, but this should be addressed in the medium term when Sommelier moves to using Ratatoskr.