Allow permissive CORS access to the metrics JSON endpoint
As this is already protected by basic auth, permissive access shouldn't be a big problem, and it is mainly to do some local debugging of a page which consumes this JSON to visualise it, and as such can be removed later.

If CORS access is generally useful, it could be added to the Sommelier routing DSL to allow pre-flight OPTIONS requests later, but it seems unnecessary for now.