Fix authentication problem with the access overview page
The last failing authentication mechanism was returning the response. In this case that meant that there was never a challenge for basic auth credentials, and access was never being authorized.

The implication is that auth mechanisms which challenge like this need to always come first, which might not be intuitive in the long run so some kind of precedence may need to be added (e.g. response 401 before 403).