Reduce noise from hack attempts in top pages display
Added a control to filter the top pages results by HTTP status code, showing only those with a 2XX status when selected. There are lots of accesses to pages like 'wp-login.php' which will 404 as they do not exist, but they are making the access log noisy. It is useful to know they are happening however, so changing the display makes more sense than filtering them at the point of recording.