Stop access log from getting too large by archiving it periodically
The table was growing too large, so this strategy is to archive old rows into CSV in another table so that the data is still accessible.

Work in progress, the archiving currently means that old data will not be available through the existing access endpoints but at least the table size will be reduced.

This is difficult to test due to race conditions and may need rethinking, but the deadline for reducing the table size is in 48 hours...